cleanup code, update example
