--- examples/router/src/switch.rs
@@ -1,1 +1,0 @@
-pub type AppAnchor = yew_router::RouterAnchor;
