+++ packages/yew-router/src/components/mod.rs
@@ -1,0 +1,2 @@
+pub mod link;
+pub mod route;
